update popup delete user
